third_party/WebKit/LayoutTests/media/media-can-play-mpeg4-video.html

<!DOCTYPE html>
<title>Test HTMLMediaElement's "canPlayType" method with unsupported video mpeg4 MIME type.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<video></video>
<script>
test(function() {
    var video = document.querySelector("video");
    // Currently no platform supports the below formats.
    assert_equals(video.canPlayType("Video/MP4; CODECS=\"mp4v.20.8, mp4a.40.2\""), "");
});
</script>
